feat(client-cli): merge `cdb-v2 download` into 'v1' command
Using the new `--backend` parameter to allow to switch to `v2` (default
to `v1`).
V2 command still exist but call 'v1' code to avoid duplication.
